Fixes from Meir's review (1)
