+++ src/forward_index.c
@@ -50,7 +50,7 @@ static size_t estimtateTermCount(const Document *doc) {
       DocumentField_GetValueCStr(field, &n);
       nChars += n;
     } else if (field->unionType == FLD_VAR_T_ARRAY) {
-        nChars += DocumentField_GetArrayValueCStrLen(field);
+        nChars += DocumentField_GetArrayValueCStrTotalLen(field);
     }
   }
   return nChars / CHARS_PER_TERM;
