+++ tests/pytests/common.py
@@ -8,7 +8,8 @@ import platform
 import itertools
 from redis.client import NEVER_DECODE
 import RLTest
-
+from typing import Any, Callable
+from RLTest.env import Query
 from includes import *
 
 
@@ -244,3 +245,30 @@ def get_redisearch_index_memory(env, index_key):
 
 def get_redisearch_vector_index_memory(env, index_key):
     return float(index_info(env, index_key)["vector_index_sz_mb"])
+
+def module_ver_filter(env, module_name, ver_filter):
+    info = env.getConnection().info()
+    for module in info['modules']:
+        if module['name'] == module_name:
+            ver = int(module['ver'])
+            return ver_filter(ver)
+    return False
+
+def has_json_api_v2(env):
+    return module_ver_filter(env, 'ReJSON', lambda ver: True if ver == 999999 else False)
+
+class ConditionalExpected:
+    
+    def __init__(self, env, cond):
+        self.env = env
+        self.cond_val = cond(env)
+        self.query = None
+
+    def call(self, *query):
+        self.query = query
+        return self
+
+    def expect_when(self, cond_val, func: Callable[[Query], Any]):
+        if cond_val == self.cond_val:
+            func(self.env.expect(*self.query))
+        return self
