+++ tests/pytests/test_json_multi_text.py
@@ -1,8 +1,7 @@
+from cmath import exp
 import json
-from typing import Any, Callable
 
 from RLTest import Env
-from RLTest.env import Query
 
 from common import *
 from includes import *
@@ -131,32 +130,6 @@ doc_non_text_content = r'''{
 }
 '''
 
-def module_ver_filter(env, module_name, ver_filter):
-    info = env.getConnection().info()
-    for module in info['modules']:
-        if module['name'] == module_name:
-            ver = int(module['ver'])
-            return ver_filter(ver)
-    return False
-
-def has_json_api_v2(env):
-    return module_ver_filter(env, 'ReJSON', lambda ver: True if ver == 999999 or ver >= 20200 else False)
-
-class ConditionalExpected:
-    
-    def __init__(self, env, cond):
-        self.env = env
-        self.cond_val = cond(env)
-        self.query = None
-
-    def call(self, *query):
-        self.query = query
-        return self
-
-    def expect_when(self, cond_val, func: Callable[[Query], Any]):
-        if cond_val == self.cond_val:
-            func(self.env.expect(*self.query))
-        return self
 
 def expect_undef_order(query : Query):
     query.error().contains("has undefined ordering")
@@ -640,6 +613,9 @@ def testconfigMultiTextOffsetDelta(env):
     env.execute_command('FT.CREATE', 'idx_category_arr', 'ON', 'JSON', 'SCHEMA', '$.category', 'AS', 'category', 'TEXT')
     waitForIndex(env, 'idx_category_arr')
 
+    env.expect('FT.CONFIG', 'SET', 'MULTI_TEXT_SLOP', '101').error().contains("Not modifiable at runtime")
+
+    
     # MULTI_TEXT_SLOP = 100 (Default)
     #
     # Offsets:
@@ -659,7 +635,7 @@ def testconfigMultiTextOffsetDelta(env):
         .expect_when(False, expect_undef_order)
     
     # MULTI_TEXT_SLOP = 101
-    env.expect('FT.CONFIG', 'SET', 'MULTI_TEXT_SLOP', '101').ok()
+    env = Env(moduleArgs = 'MULTI_TEXT_SLOP 101')
     # Offsets:
     # ["mathematics and computer science", "logic", "programming", "database"]
     #   1                2        3      ,  104   ,  205         ,  306
@@ -682,7 +658,7 @@ def testconfigMultiTextOffsetDelta(env):
         .expect_when(False, expect_undef_order)
 
     # MULTI_TEXT_SLOP = 0
-    env.expect('FT.CONFIG', 'SET', 'MULTI_TEXT_SLOP', '0').ok()
+    env = Env(moduleArgs = 'MULTI_TEXT_SLOP 0')
     # Offsets:
     # ["mathematics and computer science", "logic", "programming", "database"]
     #   1                2        3      ,  4   ,    5         ,    6
@@ -704,7 +680,15 @@ def testconfigMultiTextOffsetDelta(env):
         .expect_when(False, expect_undef_order)
 
     # MULTI_TEXT_SLOP = -1
-    env.expect('FT.CONFIG', 'SET', 'MULTI_TEXT_SLOP', '-1').error()
+    err_msg = None
+    try:
+        # Module should fail to load and should prevent redis from launching
+        env = Env(moduleArgs = 'MULTI_TEXT_SLOP -1')
+    except Exception as e:
+        if 'Cannot establish connection' in str(e):
+            err_msg = 'module init should fail due to invalid module configuration'
+
+    env.assertIsNotNone(err_msg)
 
 
 def testMultiNoHighlight(env):
